backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.5.0
pydantic-settings>=2.2.0
httpx>=0.27.0
requests>=2.31.0
numpy>=1.24.0
pillow>=10.0.0
opencv-python>=4.9.0.80
huggingface_hub>=0.24.0
